BUGFIX: Don't hang on non responsive name server.
I have experienced an bug where an name server loses its ipv6 address,
this made GetRRforZone() return an error and quit. Unfortunately the
caller in buildDNS() did non honor this because it started GetRRforZone
as an go routine and was left hanging on the channel it expected data
on.

The fix is done by making an struct that GetRRforZone returns both data
and errors on, and not to return any errors as return values.

I have also added error handling up the stream from GetRRforZone() so
the program will not halt but instead idle until next scheduled update.